memory-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.31.1
websockets>=12.0
pydantic>=2.11.7,<3.0.0
python-dateutil>=2.8.2
aiofiles>=23.2.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
